apps/backend/pyproject.toml
[tool.poetry]
name = "midjournal-app"
version = "0.1.0"
description = "Personalized RAG-powered Journal"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
fastapi = "^0.115.14"
uvicorn = "^0.35.0"
pydantic = "^2.11.7"
pydantic-settings = "^2.4.0"
langchain = "^0.3.26"
qdrant-client = "^1.9.0"
psycopg2-binary = "^2.9.10"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
bcrypt = "4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"
sqlalchemy = "^2.0.33"
pika = "^1.3.2"
sentence-transformers = "^3.0.1"
torch = "^2.3.1"
ollama = "^0.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.0"
black = "^24.4.2"
isort = "^5.13.2"
flake8 = "^7.1.0"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
